Move validation logic into CordaX500Name
