+++ core/src/main/kotlin/net/corda/core/utilities/X500NameUtils.kt
@@ -8,7 +8,6 @@ import org.bouncycastle.asn1.x500.X500NameBuilder
 import org.bouncycastle.asn1.x500.style.BCStyle
 
 val X500Name.commonName: String? get() = getRDNValueString(BCStyle.CN)
-val X500Name.organisationUnit: String? get() = getRDNValueString(BCStyle.OU)
 val X500Name.state: String? get() = getRDNValueString(BCStyle.ST)
 val X500Name.organisation: String get() = getRDNValueString(BCStyle.O) ?: throw IllegalArgumentException("Malformed X500 name, organisation attribute (O) cannot be empty.")
 val X500Name.locality: String get() = getRDNValueString(BCStyle.L) ?: throw IllegalArgumentException("Malformed X500 name, locality attribute (L) cannot be empty.")
