+++ core/src/test/kotlin/net/corda/core/identity/LegalNameValidatorTest.kt
@@ -1,7 +1,5 @@
-package net.corda.core.utilities
+package net.corda.core.identity
 
-import net.corda.core.identity.CordaX500Name
-import org.bouncycastle.asn1.x500.X500Name
 import org.junit.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
@@ -10,55 +8,55 @@ class LegalNameValidatorTest {
     @Test
     fun `no double spaces`() {
         assertFailsWith(IllegalArgumentException::class) {
-            validateLegalName("Test Legal  Name")
+            LegalNameValidator.validateLegalName("Test Legal  Name")
         }
-        validateLegalName(normaliseLegalName("Test Legal  Name"))
+        LegalNameValidator.validateLegalName(LegalNameValidator.normaliseLegalName("Test Legal  Name"))
     }
 
     @Test
     fun `no trailing white space`() {
         assertFailsWith(IllegalArgumentException::class) {
-            validateLegalName("Test ")
+            LegalNameValidator.validateLegalName("Test ")
         }
     }
 
     @Test
     fun `no prefixed white space`() {
         assertFailsWith(IllegalArgumentException::class) {
-            validateLegalName(" Test")
+            LegalNameValidator.validateLegalName(" Test")
         }
     }
 
     @Test
     fun `blacklisted words`() {
         assertFailsWith(IllegalArgumentException::class) {
-            validateLegalName("Test Server")
+            LegalNameValidator.validateLegalName("Test Server")
         }
     }
 
     @Test
     fun `blacklisted characters`() {
-        validateLegalName("Test")
+        LegalNameValidator.validateLegalName("Test")
         assertFailsWith(IllegalArgumentException::class) {
-            validateLegalName("\$Test")
+            LegalNameValidator.validateLegalName("\$Test")
         }
         assertFailsWith(IllegalArgumentException::class) {
-            validateLegalName("\"Test")
+            LegalNameValidator.validateLegalName("\"Test")
         }
         assertFailsWith(IllegalArgumentException::class) {
-            validateLegalName("\'Test")
+            LegalNameValidator.validateLegalName("\'Test")
         }
         assertFailsWith(IllegalArgumentException::class) {
-            validateLegalName("=Test")
+            LegalNameValidator.validateLegalName("=Test")
         }
     }
 
     @Test
     fun `unicode range`() {
-        validateLegalName("Test A")
+        LegalNameValidator.validateLegalName("Test A")
         assertFailsWith(IllegalArgumentException::class) {
             // Greek letter A.
-            validateLegalName("Test Α")
+            LegalNameValidator.validateLegalName("Test Α")
         }
     }
 
@@ -68,37 +66,37 @@ class LegalNameValidatorTest {
         while (longLegalName.length < 255) {
             longLegalName.append("A")
         }
-        validateLegalName(longLegalName.toString())
+        LegalNameValidator.validateLegalName(longLegalName.toString())
 
         assertFailsWith(IllegalArgumentException::class) {
-            validateLegalName(longLegalName.append("A").toString())
+            LegalNameValidator.validateLegalName(longLegalName.append("A").toString())
         }
     }
 
     @Test
     fun `legal name should be capitalized`() {
-        validateLegalName("Good legal name")
+        LegalNameValidator.validateLegalName("Good legal name")
         assertFailsWith(IllegalArgumentException::class) {
-            validateLegalName("bad name")
+            LegalNameValidator.validateLegalName("bad name")
         }
         assertFailsWith(IllegalArgumentException::class) {
-            validateLegalName("bad Name")
+            LegalNameValidator.validateLegalName("bad Name")
         }
     }
 
     @Test
     fun `correctly handle whitespaces`() {
-        assertEquals("Legal Name With Tab", normaliseLegalName("Legal Name With\tTab"))
-        assertEquals("Legal Name With Unicode Whitespaces", normaliseLegalName("Legal Name\u2004With\u0009Unicode\u0020Whitespaces"))
-        assertEquals("Legal Name With Line Breaks", normaliseLegalName("Legal Name With\n\rLine\nBreaks"))
+        assertEquals("Legal Name With Tab", LegalNameValidator.normaliseLegalName("Legal Name With\tTab"))
+        assertEquals("Legal Name With Unicode Whitespaces", LegalNameValidator.normaliseLegalName("Legal Name\u2004With\u0009Unicode\u0020Whitespaces"))
+        assertEquals("Legal Name With Line Breaks", LegalNameValidator.normaliseLegalName("Legal Name With\n\rLine\nBreaks"))
         assertFailsWith(IllegalArgumentException::class) {
-            validateLegalName("Legal Name With\tTab")
+            LegalNameValidator.validateLegalName("Legal Name With\tTab")
         }
         assertFailsWith(IllegalArgumentException::class) {
-            validateLegalName("Legal Name\u2004With\u0009Unicode\u0020Whitespaces")
+            LegalNameValidator.validateLegalName("Legal Name\u2004With\u0009Unicode\u0020Whitespaces")
         }
         assertFailsWith(IllegalArgumentException::class) {
-            validateLegalName("Legal Name With\n\rLine\nBreaks")
+            LegalNameValidator.validateLegalName("Legal Name With\n\rLine\nBreaks")
         }
     }
