+++ tools/demobench/src/main/kotlin/net/corda/demobench/model/NetworkMapConfig.kt
@@ -1,9 +1,7 @@
 package net.corda.demobench.model
 
 import net.corda.core.identity.CordaX500Name
-import net.corda.core.utilities.organisation
-import net.corda.core.utilities.WHITESPACE
-import org.bouncycastle.asn1.x500.X500Name
+import net.corda.core.identity.WHITESPACE
 
 open class NetworkMapConfig(val legalName: CordaX500Name, val p2pPort: Int) {
     val key: String = legalName.organisation.toKey()
